Improve detection of whether pages counter reference is present in the stylesheet
In case pages counter is present we don't know whether it will be used yet,
but we still create a document with relayout capability. Previously,
the check did not catch some tricky cases, now the check is performed not on the
raw CSS data but rather against the parsed CSS structures

DEVSIX-3960

Autoported commit.
Original commit hash: [0916d7d4]